+++ bootmgr-rs-core/src/system/variable.rs
@@ -5,7 +5,6 @@
 //!
 //! These store a value into a UEFI variable in a custom vendor namespace.
 
-use alloc::{vec, vec::Vec};
 use uefi::{
     CStr16, Status, guid,
     runtime::{self, VariableAttributes, VariableVendor},
@@ -16,6 +15,9 @@ use crate::{BootResult, error::BootError};
 /// The custom variable namespace for the boot manager.
 const BOOTMGR_GUID: uefi::Guid = guid!("23600d08-561e-4e68-a024-1d7d6e04ee4e");
 
+/// The maximum size of a singular type to be stored in a UEFI variable in bytes.
+const MAX_SIZE: usize = size_of::<u64>();
+
 /// A trait for implementations of UEFI variable storage.
 ///
 /// Usually this will use runtime services.
@@ -24,7 +26,6 @@ trait UefiVariableStorage {
     fn get_variable<T: UefiVariable + 'static>(
         name: &CStr16,
         vendor: &VariableVendor,
-        buf: &mut [u8],
     ) -> BootResult<T>;
 
     /// Set a variable given its name, a variable vendor, variable attributes, and the chosen type.
@@ -40,12 +41,9 @@ trait UefiVariableStorage {
 struct RuntimeUefiVariableStorage;
 
 impl UefiVariableStorage for RuntimeUefiVariableStorage {
-    fn get_variable<T: UefiVariable>(
-        name: &CStr16,
-        vendor: &VariableVendor,
-        buf: &mut [u8],
-    ) -> BootResult<T> {
-        match runtime::get_variable(name, vendor, buf) {
+    fn get_variable<T: UefiVariable>(name: &CStr16, vendor: &VariableVendor) -> BootResult<T> {
+        let mut buf = [0; MAX_SIZE];
+        match runtime::get_variable(name, vendor, &mut buf) {
             Ok((var, _)) => Ok(T::from_bytes(var)),
             Err(e) if e.status() == Status::NOT_FOUND => Ok(T::default()), // pretend that we got all zeroes if its not found
             Err(e) => Err(BootError::Uefi(e.to_err_without_payload())),
@@ -58,8 +56,17 @@ impl UefiVariableStorage for RuntimeUefiVariableStorage {
         attributes: VariableAttributes,
         num: Option<T>,
     ) -> BootResult<()> {
-        let num = num.map_or_else(|| Vec::with_capacity(0), UefiVariable::to_bytes);
-        Ok(runtime::set_variable(name, vendor, attributes, &num)?)
+        let mut bytes = 0;
+        let mut buf = [0; MAX_SIZE];
+        if let Some(num) = num {
+            bytes = num.to_bytes(&mut buf);
+        }
+        Ok(runtime::set_variable(
+            name,
+            vendor,
+            attributes,
+            &buf[0..bytes],
+        )?)
     }
 }
 
@@ -69,8 +76,11 @@ impl UefiVariableStorage for RuntimeUefiVariableStorage {
 /// are in does not particularly matter, or how these bytes are encoded or decoded, as long as the method from
 /// [`UefiVariable`] is used instead of whatever type you have. It also has to be a set size.
 pub trait UefiVariable: Sized {
-    /// Convert `Self` to a vector of bytes.
-    fn to_bytes(self) -> Vec<u8>;
+    /// The size of the type in bytes.
+    const SIZE: usize;
+
+    /// Convert `Self` to a buffer of bytes, then return the bytes written.
+    fn to_bytes(&self, out: &mut [u8]) -> usize;
 
     /// Convert a vector of bytes to `Self`.
     fn from_bytes(bytes: &[u8]) -> Self;
@@ -80,8 +90,12 @@ pub trait UefiVariable: Sized {
 }
 
 impl UefiVariable for usize {
-    fn to_bytes(self) -> Vec<u8> {
-        self.to_le_bytes().to_vec()
+    const SIZE: Self = size_of::<Self>();
+
+    fn to_bytes(&self, out: &mut [u8]) -> usize {
+        let bytes = self.to_le_bytes();
+        out[..bytes.len()].copy_from_slice(&bytes);
+        bytes.len()
     }
     fn from_bytes(bytes: &[u8]) -> Self {
         let mut array = [0; size_of::<Self>()];
@@ -94,8 +108,12 @@ impl UefiVariable for usize {
 }
 
 impl UefiVariable for u64 {
-    fn to_bytes(self) -> Vec<u8> {
-        self.to_le_bytes().to_vec()
+    const SIZE: usize = size_of::<Self>();
+
+    fn to_bytes(&self, out: &mut [u8]) -> usize {
+        let bytes = self.to_le_bytes();
+        out[..bytes.len()].copy_from_slice(&bytes);
+        bytes.len()
     }
     fn from_bytes(bytes: &[u8]) -> Self {
         let mut array = [0; size_of::<Self>()];
@@ -108,8 +126,12 @@ impl UefiVariable for u64 {
 }
 
 impl UefiVariable for u32 {
-    fn to_bytes(self) -> Vec<u8> {
-        self.to_le_bytes().to_vec()
+    const SIZE: usize = size_of::<Self>();
+
+    fn to_bytes(&self, out: &mut [u8]) -> usize {
+        let bytes = self.to_le_bytes();
+        out[..bytes.len()].copy_from_slice(&bytes);
+        bytes.len()
     }
     fn from_bytes(bytes: &[u8]) -> Self {
         let mut array = [0; size_of::<Self>()];
@@ -122,8 +144,12 @@ impl UefiVariable for u32 {
 }
 
 impl UefiVariable for u16 {
-    fn to_bytes(self) -> Vec<u8> {
-        self.to_le_bytes().to_vec()
+    const SIZE: usize = size_of::<Self>();
+
+    fn to_bytes(&self, out: &mut [u8]) -> usize {
+        let bytes = self.to_le_bytes();
+        out[..bytes.len()].copy_from_slice(&bytes);
+        bytes.len()
     }
     fn from_bytes(bytes: &[u8]) -> Self {
         let mut array = [0; size_of::<Self>()];
@@ -136,8 +162,12 @@ impl UefiVariable for u16 {
 }
 
 impl UefiVariable for u8 {
-    fn to_bytes(self) -> Vec<u8> {
-        self.to_le_bytes().to_vec()
+    const SIZE: usize = size_of::<Self>();
+
+    fn to_bytes(&self, out: &mut [u8]) -> usize {
+        let bytes = self.to_le_bytes();
+        out[..bytes.len()].copy_from_slice(&bytes);
+        bytes.len()
     }
     fn from_bytes(bytes: &[u8]) -> Self {
         let mut array = [0; size_of::<Self>()];
@@ -150,8 +180,12 @@ impl UefiVariable for u8 {
 }
 
 impl UefiVariable for bool {
-    fn to_bytes(self) -> Vec<u8> {
-        u8::from(self).to_le_bytes().to_vec()
+    const SIZE: usize = size_of::<Self>();
+
+    fn to_bytes(&self, out: &mut [u8]) -> usize {
+        let bytes = u8::from(*self).to_le_bytes();
+        out[..bytes.len()].copy_from_slice(&bytes);
+        bytes.len()
     }
     fn from_bytes(bytes: &[u8]) -> Self {
         let mut array = [0; size_of::<Self>()];
@@ -208,7 +242,6 @@ pub fn get_variable<T: UefiVariable + 'static>(
     name: &CStr16,
     vendor: Option<VariableVendor>,
 ) -> BootResult<T> {
-    let mut buf = vec![Default::default(); size_of::<T>()];
     let vendor = vendor.unwrap_or(runtime::VariableVendor(BOOTMGR_GUID));
-    RuntimeUefiVariableStorage::get_variable(name, &vendor, &mut buf)
+    RuntimeUefiVariableStorage::get_variable(name, &vendor)
 }